[refactor] component and styles architecture -> page based and shared
