+++ .gitignore
@@ -1,3 +1,5 @@
+*.vs
+
 *.env
 .env*
