+++ frontend/src/components/Navbar.tsx
@@ -1,6 +1,6 @@
 import { Link } from "react-router-dom";
-import { ModeToggle } from "./ui/theme-mode-toggle";
-import { useTheme } from "./ui/theme-provider";
+import { ModeToggle } from "./ui/theme/theme-mode-toggle";
+import { useTheme } from "./ui/theme/theme-provider";
 import logoLight from "../assets/logos/logoLight.svg";
 import logoDark from "../assets/logos/logoDark.svg";
 import "../styles/globals.css"
