+++ frontend/src/components/ui/animated-list/animated-list.tsx
@@ -6,7 +6,7 @@ import React, {
 } from "react"
 import { AnimatePresence, motion, type MotionProps } from "motion/react"
 
-import { cn } from "../../lib/utils"
+import { cn } from "../../../lib/utils"
 
 export function AnimatedListItem({ children }: { children: React.ReactNode }) {
   const animations: MotionProps = {
