+++ frontend/src/pages/home-page/components/HeroCardUI.tsx
