+++ frontend/src/pages/home-page/components/HighlightCards.tsx
@@ -1,4 +1,4 @@
-import { Card, CardContent, CardTitle } from "../components/ui/card";
+import { Card, CardContent, CardTitle } from "../../../components/ui/card";
 
 export default function Features() {
   return (
