+++ frontend/src/pages/portfolio-page/PortfolioList.tsx
@@ -1,9 +1,9 @@
 import React, { useEffect, useState } from "react";
-import { getPortfolios, deletePortfolio } from "../api/api";
-import type { Portfolio } from "../types/portfolio";
+import { getPortfolios, deletePortfolio } from "../../api/api";
+import type { Portfolio } from "../../types/portfolio";
 import { Link } from "react-router-dom";
-import AlertPopUp  from "../components/ui/alert";
-import { Button } from "../components/ui/button";
+import AlertPopUp  from "../../components/ui/alert";
+import { Button } from "../../components/ui/button";
 import { AnimatePresence } from "framer-motion";
 
 type AlertColor =
