+++ frontend/src/pages/portfolio-page/components/FileUpload.tsx
@@ -1,7 +1,7 @@
 import { useState } from "react";
 import axios from "axios";
 
-import { Button } from "./ui/button";
+import { Button } from "../../../components/ui/button";
 import { Spinner } from "@heroui/spinner";
 
 const FileUpload = () => {
