+++ frontend/src/pages/risk-page/RiskResultPage.tsx
@@ -1,10 +1,10 @@
 import React, { useEffect, useState } from "react";
-import { getRiskStatus } from "../api/api";
-import type { RiskResult } from "../types/risk";
-import RiskChart from "../components/RiskChart";
+import { getRiskStatus } from "../../api/api";
+import type { RiskResult } from "../../types/risk";
+import RiskChart from "./components/RiskChart";
 import { useParams } from "react-router-dom";
 import {Spinner} from "@heroui/react";
-import AlertPopUp  from "../components/ui/alert";
+import AlertPopUp  from "../../components/ui/alert";
 import { AnimatePresence } from "framer-motion";
 
 type AlertColor =
