+++ frontend/src/pages/risk-page/components/RiskChart.tsx
