+++ frontend/src/pages/risk-page/components/RiskTable.tsx
